spawnerEditor adjustements + new health regeneration functionality
Some improvements to spawnerEditor script and new Health regeneration script that allows health to regenerate.  Can be only added to a gameobject with Health component

+++ Assets/Scripts/UI/UIHealthBar.cs
@@ -27,7 +27,7 @@ namespace Scripts.GamePlay.UI
 
         private void OnCurrentHealthValueChanged()
         {
-            healthBar.value = (float) health.CurrentHealth / health.MaxHealth;
+            healthBar.value = health.CurrentHealth / health.MaxHealth;
         }
     }
 }
